JSVALUE32_64 DFG JIT - fillJSValue with a pair of GPRs should not set the registerFormat to be DataFormatJSDouble https://bugs.webkit.org/show_bug.cgi?id=69720
Patch by Yuqiang Xian <yuqiang.xian@intel.com> on 2011-10-09
Reviewed by Filip Pizlo.

In JSVALUE32_64 DFG, DataFormatJSDouble is assumed to be represented by
a FPR and will be used for further optimizations, though we currently
don't fully utilize it. For now when filling a JS value which was
spilled as a JSDouble with a pair of GPRs, we'll set the registerFormat
to DataFormatJS to avoid compilation errors.

* dfg/DFGJITCodeGenerator32_64.cpp:
(JSC::DFG::JITCodeGenerator::fillJSValue):

Identifier: 85776@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97031 268f45cc-cd09-0410-ab3c-d52691b4dbfc