[EFL] DRT: Do not abort() when an unknown key is pressed. https://bugs.webkit.org/show_bug.cgi?id=69856
Patch by Raphael Kubo da Costa <kubo@profusion.mobi> on 2011-10-11
Reviewed by Antonio Gomes.

A lot of tests crash due to the previous behaviour. Just using the
received name of the pressed key works most of the time, and when it
does not the failing tests should tell us.

* DumpRenderTree/efl/EventSender.cpp:
(keyPadNameFromJSValue):
(keyNameFromJSValue):
(keyDownCallback):

Identifier: 85903@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97169 268f45cc-cd09-0410-ab3c-d52691b4dbfc