Hang beneath PluginProcessProxy::createPropertyListFile() when using a thin build https://bugs.webkit.org/show_bug.cgi?id=69960
Reviewed by Adam Roben.

* UIProcess/Plugins/mac/PluginProcessProxyMac.mm:
(WebKit::PluginProcessProxy::createPropertyListFile): Check for any non-zero return value
from posix_spawn() rather than for a negative return value. The old, incorrect check, caused
this function to hang in waitpid() after failing to spawn a plug-in process for a 32-bit plug-in
when WebKit2 was built for 64-bit only.



Identifier: 86035@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97304 268f45cc-cd09-0410-ab3c-d52691b4dbfc