Source/WebCore: https://bugs.webkit.org/show_bug.cgi?id=69932
Layout repainting messed up for objects in regions. There were two bugs here.
The first is that the regionOverflowRect needs to at least be the size of the
region's rect. When you're in the middle of layout, you haven't computed the
overflow rect for the flow thread yet (the flow thread's height could even be 0),
and so you just need to enforce a minimum size. Also removed the region-overflow:break
clipping behavior, since it was established that was wrong.

The second bug is that when you tell a region to repaint, you need to disable layout state,
since applying layout deltas from the flow thread makes no sense. The region is in a totally
different place in the render tree, so the current flow thread layout state can't be used
to determine your coordinate offset.

Reviewed by Dan Bernstein.

Added fast/repaint test.

* rendering/RenderFlowThread.cpp:
(WebCore::RenderFlowThread::repaintRectangleInRegions):
* rendering/RenderRegion.cpp:
(WebCore::RenderRegion::regionOverflowRect):

LayoutTests: https://bugs.webkit.org/show_bug.cgi?id=69932

Layout repainting messed up for objects in regions. Add a repaint test to
illustrate the issue.

Reviewed by Dan Bernstein.

* fast/repaint/region-painting-via-layout.html: Added.
* platform/mac/fast/regions/region-overflow-break-expected.png:
* platform/mac/fast/repaint/region-painting-via-layout-expected.png: Added.
* platform/mac/fast/repaint/region-painting-via-layout-expected.txt: Added.



Identifier: 86100@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97370 268f45cc-cd09-0410-ab3c-d52691b4dbfc