[Qt] Remove dead code in build-webkit
build-webkit will iterate over a list of subdirectories to build,
but this feature is only used by the Apple ports. For QtWebKit we
always pass an empty directory to buildQMakeProject, which makes
the logic for dealing with sub-projects in buildQMakeProject just
added noise.

If we do want this feature at some point, we should redo it in
the context of the upcoming build system changes for Qt.

Reviewed by Simon Hausmann.

Identifier: 86456@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97735 268f45cc-cd09-0410-ab3c-d52691b4dbfc