Crash in RenderDeprecatedFlexibleBox::layoutHorizontalBox https://bugs.webkit.org/show_bug.cgi?id=70183
Patch by Julien Chaffraix <jchaffraix@webkit.org> on 2011-10-18
Reviewed by David Hyatt.

Source/WebCore:

Test: fast/flexbox/layoutHorizontal-crash.html

The deferred scroll information update logic was not updated when an object was destroy'ed.

The fix is very simple. As the deferred logic was only used with deprecated flexbox, it is a
pretty uncommon cases so let's optimize the new code for the common case.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::willBeDestroyed):

LayoutTests:

* fast/flexbox/layoutHorizontal-crash-expected.txt: Added.
* fast/flexbox/layoutHorizontal-crash.html: Added.


Identifier: 86504@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97786 268f45cc-cd09-0410-ab3c-d52691b4dbfc