Crash in WebProcess at com.apple.AppKit: -[NSFilePromiseDragSource draggedImage:endedAt:operation:] + 101 https://bugs.webkit.org/show_bug.cgi?id=70340 <rdar://problem/9548174>
We have seen a number of crash reports for this issue that seem to occurr when
the user drags an image from a web page onto the desktop.
We have not been able to reproduce this crash, therefore this is a speculative fix,
but the theory is that the dragging source has been already released when draggedImage
is called.
This patch attempts a fix for the problem by retaining the dragging source when
NSFilePromiseDragSource is created.

Reviewed by Darin Adler.

* WebProcess/WebCoreSupport/mac/WebDragClientMac.mm:
(-[WKPasteboardFilePromiseOwner initWithSource:]):
(-[WKPasteboardFilePromiseOwner dealloc]):



Identifier: 86525@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97808 268f45cc-cd09-0410-ab3c-d52691b4dbfc