[WK2] WebFrameLoaderClient::shouldFallback() should use a port-specific implementation https://bugs.webkit.org/show_bug.cgi?id=70055
Patch by Jesus Sanchez-Palencia <jesus.palencia@openbossa.org> on 2011-10-21
Reviewed by Luiz Agostini.

In QtWebKit 1, our FrameLoaderClient::shouldFallback was checking for InterruptedForPolicyChange
error. In WebKit2, we are relying on the common implementation for WebFrameLoaderClient::shouldFallback
which doesn't check for the same error. Here we add this check only for Qt.

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::shouldFallBack):

Identifier: 86834@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98138 268f45cc-cd09-0410-ab3c-d52691b4dbfc