Make relayout-nested-positioned-elements-crash.html more reliable https://bugs.webkit.org/show_bug.cgi?id=70672
Reviewed by Kenneth Rohde Christiansen.

The test relayout-nested-positioned-elements-crash relies on implementation
details of the test runner. If the function runTest() is run
during the test, the result will be different than the expected results.

Change the test to use the Javascript test framework.
Ensure the element has the correct style in the final state.

* fast/block/positioning/relayout-nested-positioned-elements-crash-expected.txt:
* fast/block/positioning/relayout-nested-positioned-elements-crash.html:


Identifier: 86874@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98197 268f45cc-cd09-0410-ab3c-d52691b4dbfc