selectedIndex gets set from -1 to 0 when modifying options https://bugs.webkit.org/show_bug.cgi?id=70547 <rdar://problem/8388856>
Reviewed by Darin Adler.

Source/WebCore:

Changing the text causes a recalculation of the list items, which in the menu list case
forces the first element to be selected. We check the value of the selected option prior,
and restore it if it differs.

Test: fast/dom/HTMLSelectElement/selected-index-preserved-when-option-text-changes.html

* html/HTMLOptionElement.cpp:
(WebCore::HTMLOptionElement::setText):
* html/HTMLSelectElement.h: promote usesMenuList() from private to public for use by HTMLOptionElement

LayoutTests:

The added tests set the selected index to either -1 or 1 (a legitimate value), and then change
the text, value, and label of the option. In all cases, the selected index should remain unchanged.

* fast/dom/HTMLSelectElement/selected-index-preserved-when-option-text-changes-expected.txt: Added.
* fast/dom/HTMLSelectElement/selected-index-preserved-when-option-text-changes.html: Added.

Identifier: 87155@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98505 268f45cc-cd09-0410-ab3c-d52691b4dbfc