Source/JavaScriptCore: Object.getOwnPropertyDescriptor() does not retrieve the getter/setter from a property on the window that has been overridden with a getter/setter https://bugs.webkit.org/show_bug.cgi?id=71333
Reviewed by Gavin Barraclough.

Tested by fast/dom/getter-on-window-object2.html

* runtime/PropertyDescriptor.cpp:
(JSC::PropertyDescriptor::setDescriptor):
The attributes returned from Structure::get do not include Getter or Setter, so
instead check if the value is a GetterSetter like we do elsewhere. If it is, update
the descriptor's attributes accordingly.

LayoutTests: Object.getOwnPropertyDescriptor() does not retrieve the getter/setter from a property on the window that has been overridden with a getter/setter/
https://bugs.webkit.org/show_bug.cgi?id=71333

Reviewed by Gavin Barraclough.

* fast/dom/getter-on-window-object2-expected.txt:
Update for now correct results.


Identifier: 87727@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99136 268f45cc-cd09-0410-ab3c-d52691b4dbfc