Remove deprecated free functions in port.factory https://bugs.webkit.org/show_bug.cgi?id=71494
Unreviewed.  Fixing exception seen when running test-webkitpy.

Sorry, I previously had deleted the relevant rebaseline-chromium-webkit-tests
unittest, since that script is nearly ready to be deleted.  But I added
it back at the last second and failed to run the tests. :(
This fixes the exception the bots were seeing.

* Scripts/webkitpy/style/checkers/test_expectations.py:
* Scripts/webkitpy/to_be_moved/rebaseline_chromium_webkit_tests.py:
* Scripts/webkitpy/to_be_moved/rebaseline_chromium_webkit_tests_unittest.py:

Identifier: 87858@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99271 268f45cc-cd09-0410-ab3c-d52691b4dbfc